no-issue: Filter matches which supports complex filter conditions with all possible combinations
- Experimented with spring-data-rest + QuerydslPredicateExecutor, QuerydslBinderCustomizer, Predicate
  - This is an excellent way to build dynamic query/search web application
  - Drawback is, no explicit way to integrate GeoSpatial queries

- QueryBuilder: Exposes fluent API that simplifies creation of Query
- MatchService: Leverages MongoTemplate to achieve complex filter conditions with all possible combinations

+++ src/main/java/com/dating/matcher/domain/MatchDocument.java
@@ -17,13 +17,13 @@ public class MatchDocument {
 
     private String jobTitle;
 
-    private float height;
+    private double height;
 
     private CityDocument city;
 
     private String mainPhoto;
 
-    private float compatibilityScore;
+    private double compatibilityScore;
 
     private int contactsExchanged;
